Use int instead of short for local index.

+++ MiniJavaCompiler/Support/AbstractSyntaxTree/VariableDeclaration.cs
@@ -13,11 +13,11 @@ namespace MiniJavaCompiler.Support.AbstractSyntaxTree
             Class
         }
         public Kind VariableKind { get; private set; }
-        public short LocalIndex { get; set; }
+        public int LocalIndex { get; set; }
         public bool IsInitialized { get; set; }
 
         public VariableDeclaration(string name, string type, bool isArray,
-            Kind kind, short localIndex, int row, int col)
+            Kind kind, int localIndex, int row, int col)
             : base(name, type, isArray, row, col)
         {
             VariableKind = kind;
